Add three stages of calculation
Each part of the TFT calculator gets increasingly advanced, adding three units to the team.  The framework is the same for each, but reads off of a different file to compile a list of teams.  With each team made, the algorithm then assigns it three scores, one for affiliations, one for links, and one for buffs.

Affiliations are how many connections to other units through class similarities the team has as a whole.
Links are how many class connections are present within the team.
Buffs are how many levels of buff have been achieved with the amount of links in the team.

In the first two calculations, the number of possible teams is limited enough to warrant sorting by links and buffs.
However, the last calculation has upwards of 19 million possible teams.
For the last calculation, to avoid memory problems, team data is outputted to a number of files, which may then be read off of for further manipulation.